server/requirements.txt
fastapi[standard]>=0.109.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
httpx>=0.25.0
python-dotenv>=1.0.0
